relocate logging interface to the reusable library location
